web/pokemon-strength-r99bbit/requirements.txt
pyyaml==5.3.0
Flask
flask_cors